requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
numpy==1.23.5
tqdm==4.66.1
scikit-learn==1.2.2
kaggle==1.5.16
pillow==10.0.0 
